fix: data structure of widget, use value instead of @id
